Voting system: javascript major changes
Thanks to some javascript major changes, we have a Reddit-style voting
system now. Just two buttons are provided: up vote and down vote.

At first buttons are grey, so to indicate the user has not placed any
vote yet. As soon as the user votes, the corresponding icon gets
coloured (green for upvote, red for downvote).

On a second press of the same button, the grey colour is restored and
the vote is cleared. No need of a specific "clear vote" button then.

When the page is loaded again, every voting button is properly
coloured as for the user past choice.

Still to do:
- code cleaning (that's never enough)
- upvote and downvote icons will be replaced by web designers