Now bookmarking application is more generic.
Note: the previous implementation was BROKEN, since it manipulated ``Act``
instances, not instances of concrete ``Act`` subclasses (deliberations, etc.),
as it should.

The simpler approach I was able to think about to circumvent this problem was to
rely on Django's content type framework to target a specific object.

This way, we also have the added benefit of quasi-generic app that might be used
to bookmark any model declaring a ``BooleanField`` named ``is_key``.