Events app major refactoring.
Events are shown in templates both as lists and as calendar
widgets. When showing a list of events, we only want future ones. When
showing a calendar widget we want future events and past ones back to
N days in the past. At the moment N = 31 (aka one month), but that can
be customized (see ``EVENTS_DAYS_IN_THE_PAST``).

Usage:

* in a view:
    extra_context = {
        'events' = Event.future.filter(institution__institution_type=Institution.COUNCIL),
        }

* in a template, list of events (only future events):
    {% for event in events %}
    {% if not event.is_past_due %}
    <li><strong>{{ event.date|date:"d.m.Y" }}</strong> <a href="{{ event.act.downcast.get_absolute_url }}">{{ event.act.title }}</a></li>
    {% endif %}
    {% endfor %}

* in a template, calendar widget (future events and past ones back to N days ago):
    {% for event in events %}
    <li data-day="{{ event.date|date:"d/m/Y" }}"><a href="{{ event.act.downcast.get_absolute_url }}">{{ event.act.title }}</a></li>
    {% endfor %}