A present/absent graph for politician detail page.
Politician detail pages must provide a "presence" graph when
politician is a counselor. This graph shows how many times the
counselor was present during votations and how many was absent.

Average absence/presence values are provided as well.

Please, check out the algorithm used for such calculations and see if
it complies with the wanted behaviour.