Politician's acts - a first draft
This is a first implementation of the list of acts signed or co-signed by a politician.
The "act's status" column needs to be revised and maybe improved.
The "users' votes" and "comments" columns work, but need to be tested because there are no related data in DB now.
The number of each single act's comments in the "comments" column needs to be actually linked... to what, I don't know.