added transition states for deliberations
Added special transition states for elements
CouncilDeliberation and CityGovernmentDeliberation. the new
states express the fact that such deliberations are:
- discussed in a commission
- accepted by a commission
- rejected by a commission

since the same act can be analyzed by more than a commission,
such elements can appear multiple times in the same act, and we
expect that the @note attribute of CouncilDeliberationTransition
(resp. CityGovernmentDeliberationTransition) the name of the
specif commission appears.

close #151