refactored import for ballots from mdb files
now when mapping an external ID to an internal object, Many-to-One
mappings are allowed, to handle the case of a counselor that replaces
another counselor and continue to use its voting card, in the next
ballots. 

introduced an integrity check: when an external id is mapped onto an
OM charge, check that the latter was in charge at that day. also, in
case many OM charges are mapped onto the same external id, check than
only one was in charge at that day.

see #368