fix #566
note that the correct choices can be shown only when editing an act. on the
contrary, when adding an act we don't know the act type, thus we don't know
what transition STATUS to select. in that case we show Act.STATUS options.
in the future we may think to show all STATUS options from Act subclasses
(to be as generic as possible). let's think about it ...